refactor: moving the mariadb client
